// src/app/core/models/registro-request.ts
export interface RegistroRequest {
  idTipoDocIdentidad: number;
  idTipoPersona: number;
  codUbigeo: string;
  nroDocumento: string;
  nombres: string;
  apellidoPaterno: string;
  apellidoMaterno: string;
  direccion: string;
  correo: string;
  celular: string;
  ruc: string;
  razonSocial: string;
  username: string;
  password: string;
}